Move cddl related things into own mod
