+++ p2panda-rs/src/cddl/validation.rs
@@ -2,13 +2,13 @@
 
 use cddl::validator::cbor;
 
-use super::SchemaValidationError;
+use crate::cddl::CDDLValidationError;
 
-/// Checks CBOR bytes against CDDL schemas.
+/// Checks CBOR bytes against CDDL.
 ///
 /// This method also converts validation errors coming from the `cddl` crate into an
 /// concatenated error operation and returns it.
-pub fn validate_schema(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), SchemaValidationError> {
+pub fn validate_cddl(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), CDDLValidationError> {
     match cddl::validate_cbor_from_slice(cddl_schema, &bytes) {
         Err(cbor::Error::Validation(err)) => {
             let err_str = err
@@ -23,10 +23,10 @@ pub fn validate_schema(cddl_schema: &str, bytes: Vec<u8>) -> Result<(), SchemaVa
                 })
                 .collect::<Vec<String>>();
 
-            Err(SchemaValidationError::InvalidSchema(err_str))
+            Err(CDDLValidationError::InvalidSchema(err_str))
         }
-        Err(cbor::Error::CBORParsing(_err)) => Err(SchemaValidationError::InvalidCBOR),
-        Err(cbor::Error::CDDLParsing(err)) => Err(SchemaValidationError::InvalidCDDL(err)),
+        Err(cbor::Error::CBORParsing(_err)) => Err(CDDLValidationError::InvalidCBOR),
+        Err(cbor::Error::CDDLParsing(err)) => Err(CDDLValidationError::InvalidCDDL(err)),
         _ => Ok(()),
     }
 }
@@ -36,15 +36,15 @@ mod tests {
     use ciborium::cbor;
     use rstest::rstest;
 
-    use crate::{
-        operation::OperationEncoded,
-        schema::{validation::validate_schema, OPERATION_SCHEMA},
-        test_utils::fixtures::operation_encoded,
-    };
+    use crate::cddl::OPERATION_FORMAT;
+    use crate::operation::OperationEncoded;
+    use crate::test_utils::fixtures::operation_encoded;
+
+    use super::validate_cddl;
 
     #[rstest]
     fn validate_operation_cbor(operation_encoded: OperationEncoded) {
-        assert!(validate_schema(OPERATION_SCHEMA, operation_encoded.to_bytes()).is_ok())
+        assert!(validate_cddl(OPERATION_FORMAT, operation_encoded.to_bytes()).is_ok())
     }
 
     #[test]
@@ -56,15 +56,17 @@ mod tests {
         }
         "#;
 
+        let age: usize = 4;
+
         let value = cbor!({
             "name" => "Latte",
-            "age" => 4
+            "age" => age,
         })
         .unwrap();
 
         let mut cbor_bytes = Vec::new();
         ciborium::ser::into_writer(&value, &mut cbor_bytes).unwrap();
-        assert!(validate_schema(cddl, cbor_bytes).is_ok());
+        assert!(validate_cddl(cddl, cbor_bytes).is_ok());
     }
 
     #[test]
@@ -84,7 +86,7 @@ mod tests {
 
         let mut cbor_bytes = Vec::new();
         ciborium::ser::into_writer(&value, &mut cbor_bytes).unwrap();
-        assert!(validate_schema(cddl, cbor_bytes).is_err());
+        assert!(validate_cddl(cddl, cbor_bytes).is_err());
     }
 
     #[test]
@@ -96,9 +98,9 @@ mod tests {
         }
         "#;
 
-        // invalid CBOR
+        // Invalid CBOR
         let cbor_bytes = Vec::from("}");
-        assert!(validate_schema(cddl, cbor_bytes).is_err());
+        assert!(validate_cddl(cddl, cbor_bytes).is_err());
     }
 
     #[test]
@@ -110,6 +112,6 @@ mod tests {
         }
         "#;
 
-        assert!(validate_schema(cddl, vec![1u8]).is_err());
+        assert!(validate_cddl(cddl, vec![1u8]).is_err());
     }
 }
